+++ game/board.py
@@ -8,6 +8,10 @@ DOUBLE_WORD_SCORE = ((1,1), (2,2), (3,3), (4,4), (1, 13), (2, 12), (3, 11), (4, 
 TRIPLE_LETTER_SCORE = ((1,5), (1, 9), (5,1), (5,5), (5,9), (5,13), (9,1), (9,5), (9,9), (9,13), (13, 5), (13,9))
 DOUBLE_LETTER_SCORE = ((0, 3), (0,11), (2,6), (2,8), (3,0), (3,7), (3,14), (6,2), (6,6), (6,8), (6,12), (7,3), (7,11),
                         (8,2), (8,6), (8,8), (8, 12), (11,0), (11,7), (11,14), (12,6), (12,8), (14, 3), (14, 11))
+values = {
+    'A': 1, 'B': 3, 'C': 3, 'D': 2, 'E': 1, 'F': 4, 'G': 2, 'H': 4, 'I': 1, 'J': 8, 'L': 1, 'M': 3, 'N': 1, 'Ñ': 8, 'O': 1,
+    'P': 3, 'Q': 5, 'R': 1, 'S': 1, 'T': 1, 'U': 1, 'V': 4, 'X': 8, 'Y': 4, 'Z': 10, '*':0
+}
 
 class SoloVoHParaLaOrientacion(Exception):
     pass
@@ -44,10 +48,6 @@ class Board:
             self.set_cell_multiplier(coordinate, "letter", 3)
         for coordinate in DOUBLE_LETTER_SCORE:
             self.set_cell_multiplier(coordinate, "letter", 2)
-    
-    def deactivate_cell(self, row, col):
-        self.grid[row][col].multiplier = 1
-        self.grid[row][col].multiplier_type = ""
 
     def validate_word_horizontal(self, word, location):
         x, y = location
@@ -62,9 +62,9 @@ class Board:
         return True
 
     def validate_word_inside_board(self, word, location, orientation):
-        if orientation.lower() == "h":
+        if orientation.upper() == "H":
             return self.validate_word_horizontal(word, location)
-        elif orientation.lower() == "v":
+        elif orientation.upper() == "V":
             return self.validate_word_vertical(word, location)
         else:
             raise SoloVoHParaLaOrientacion(Exception)
@@ -115,16 +115,58 @@ class Board:
         
     def put_word(self, word, location, orientation):
         x, y = location
-        if orientation.lower() == 'h':
-            for i, tile in enumerate(word):
-                self.grid[x][y + i].add_letter(tile)
-        elif orientation.lower() == 'v':
-            for i, tile in enumerate(word):
-                self.grid[x + i][y].add_letter(tile)
-        else:
-            raise SoloVoHParaLaOrientacion(Exception)
-        
+        cells = []
+        for i, letter in enumerate(word):
+            if orientation.upper() == 'H':
+                cell= self.grid[x][y + i]
+            elif orientation.upper() == 'V':
+                cell=self.grid[x + i][y]
+            
+            value = values.get(letter, 0)
+            tile = Tile(letter=letter, value=value)
+            cell.add_letter(tile)
+            cells.append(cell)
+        return cells
+            
+    def get_word_cells(self, word, location, orientation):
+        word_cells = [] 
+        row, col = location 
 
+        for letter in word:
+            word_cells.append(self.grid[row][col])  
+            if orientation.upper() == 'H':
+                col += 1  
+            elif orientation.upper()=='V':
+                row += 1  
+        return word_cells
+    
+    @staticmethod
+    def calculate_word_value(word: list[Cell]) -> int:
+        value: int = 0
+        multiplier_word = None
+
+        for cell in word:
+            value = value + cell.calculate_value()
+            if cell.multiplier_type == "word" and cell.active:
+                multiplier_word = cell.multiplier
+                cell.active == False
+        if multiplier_word:
+            value = value * multiplier_word
+        return value
+                
+    # def is_valid_crossword(self, word, location, orientation):
+    #     x, y = location
+    #     for i, letter in enumerate(word):
+    #         if orientation.lower() == 'h':
+    #             if (y + i - 1 >= 0 and self.grid[x][y + i - 1].letter is not None) or \
+    #                (y + i + 1 < 15 and self.grid[x][y + i + 1].letter is not None):
+    #                 return True
+    #         elif orientation.lower() == 'v':
+    #             if (x + i - 1 >= 0 and self.grid[x + i - 1][y].letter is not None) or \
+    #                (x + i + 1 < 15 and self.grid[x + i + 1][y].letter is not None):
+    #                 return True
+
+    #     return False