+++ game/player.py
@@ -4,23 +4,28 @@ from game.tiles import Tile
 class Player:
 
     def __init__(self, bag_tiles=BagTiles()):
+        self.bag_tiles = BagTiles() 
         self.tiles = bag_tiles.take(7)
-        self.bag_tiles = bag_tiles
         self.score = 0
     
     def rellenar(self):
         self.tiles += self.bag_tiles.take(7 - len(self.tiles))
 
     def has_letters(self, word):
-        word_tiles_counts = {letter: word.count(letter) for letter in word}
-        player_tiles_counts = {tile.letter: self.tiles.count(tile) for tile in self.tiles}
-
-        for letter, count in word_tiles_counts.items():
-            if letter not in player_tiles_counts or count > player_tiles_counts[letter]:
+        player_tiles = self.tiles.copy()
+        word_letters_counts = {letter: word.count(letter) for letter in word}
+        
+        for letter, count in word_letters_counts.items():
+            found_tiles = [tile for tile in player_tiles if tile.letter == letter]
+            if len(found_tiles) < count:
                 return False
+            
+            for _ in range(count):
+                player_tiles.remove(found_tiles.pop(0))
+        
+        self.tiles = player_tiles
+        
         return True
     
-    # def __str__(self):
-    #     return f"{self.player_index}"
     
     #def take_tile_using_letters()